requirements.txt
web3>=6.0.0
python-dotenv>=1.0.0
eth-typing>=3.0.0
eth-account>=0.8.0
typing-extensions>=4.5.0
pytest>=7.0.0
